Move `communicate` from util to linter
Keep a bare-bone `communicate` in util and mark it deprecated.
