+++ lint/util.py
@@ -1,7 +1,5 @@
 """This module provides general utility methods."""
 
-from collections import ChainMap
-from contextlib import contextmanager
 from functools import lru_cache
 import locale
 import logging
@@ -188,56 +186,7 @@ def check_output(cmd, cwd=None):
         return process_popen_output(output)
 
 
-RUNNING_TEMPLATE = """{headline}
-
-  {cwd}  (working dir)
-  {prompt}{pipe} {cmd}
-"""
-
-PIPE_TEMPLATE = ' type {} |' if os.name == 'nt' else ' cat {} |'
-ENV_TEMPLATE = """
-  Modified environment:
-
-  {env}
-
-  Type: `import os; os.environ` in the Sublime console to get the full environment.
-"""
-
-
-def make_nice_log_message(headline, cmd, is_stdin,
-                          cwd=None, view=None, env=None):
-    import pprint
-    import textwrap
-
-    filename = view.file_name() if view else None
-    if filename and cwd:
-        rel_filename = os.path.relpath(filename, cwd)
-    elif not filename:
-        rel_filename = '<buffer {}>'.format(view.buffer_id()) if view else '<?>'
-
-    real_cwd = cwd if cwd else os.path.realpath(os.path.curdir)
-
-    exec_msg = RUNNING_TEMPLATE.format(
-        headline=headline,
-        cwd=real_cwd,
-        prompt='>' if os.name == 'nt' else '$',
-        pipe=PIPE_TEMPLATE.format(rel_filename) if is_stdin else '',
-        cmd=' '.join(cmd)
-    )
-
-    env_msg = ENV_TEMPLATE.format(
-        env=textwrap.indent(
-            pprint.pformat(env, indent=2),
-            '  ',
-            predicate=lambda line: not line.startswith('{')
-        )
-    ) if env else ''
-
-    return exec_msg + env_msg
-
-
-def communicate(cmd, code=None, output_stream=STREAM_STDOUT, env=None, cwd=None,
-                _linter=None):
+def communicate(cmd, code=None, output_stream=STREAM_STDOUT, env=None, cwd=None):
     """
     Return the result of sending code via stdin to an executable.
 
@@ -246,13 +195,13 @@ def communicate(cmd, code=None, output_stream=STREAM_STDOUT, env=None, cwd=None,
     If env is None, the result of create_environment is used.
 
     """
+    logger.warning('`util.communicate` has been deprecated.')
+
     if code is not None:
         code = code.encode('utf8')
     if env is None:
         env = create_environment()
 
-    view = _linter.view if _linter else None
-    bid = view.buffer_id() if view else None
     uses_stdin = code is not None
 
     try:
@@ -261,57 +210,22 @@ def communicate(cmd, code=None, output_stream=STREAM_STDOUT, env=None, cwd=None,
             stdin=subprocess.PIPE if uses_stdin else None,
             stdout=subprocess.PIPE if output_stream & STREAM_STDOUT else None,
             stderr=subprocess.PIPE if output_stream & STREAM_STDERR else None,
-            startupinfo=create_startupinfo(),
-            creationflags=get_creationflags()
+            startupinfo=create_startupinfo()
         )
     except Exception as err:
-        try:
-            augmented_env = dict(ChainMap(*env.maps[0:-1]))
-        except AttributeError:
-            augmented_env = None
-        logger.error(make_nice_log_message(
-            '  Execution failed\n\n  {}'.format(str(err)),
-            cmd, uses_stdin, cwd, view, augmented_env))
-
-        if _linter:
-            _linter.notify_failure()
+        logger.error('  Execution failed\n\n  {}\n  {}{}'.format(
+            str(err),
+            ' '.join(cmd),
+            '\n  {}'.format(cwd) if cwd else ''
+        ))
 
         return ''
 
     if logger.isEnabledFor(logging.INFO):
-        logger.info(make_nice_log_message(
-            'Running ...', cmd, uses_stdin, cwd, view, None))
+        logger.info('Running `{}`'.format(' '.join(cmd)))
 
-    with store_proc_while_running(bid, proc):
-        try:
-            out = proc.communicate(code)
-        except BrokenPipeError as err:
-            friendly_terminated = getattr(proc, 'friendly_terminated', False)
-            # TODO: Consider `raise err from None` for friendly_terminated
-            # Bc this is a somewhat expected failure which should NOT count
-            # as 'no errors'. E.g. it should NOT empty all highlighted regions.
-            if friendly_terminated:
-                logger.warning('Broken pipe in <pid {}>'.format(proc.pid))
-            else:
-                # TODO: Consider `notify_failure()` here
-                logger.warning('Exception: {}'.format(str(err)))
-            return ''
-
+    out = proc.communicate(code)
     return popen_output(proc, *out)
-
-
-@contextmanager
-def store_proc_while_running(bid, proc):
-    from . import persist
-
-    with persist.active_procs_lock:
-        persist.active_procs[bid].append(proc)
-
-    try:
-        yield proc
-    finally:
-        with persist.active_procs_lock:
-            persist.active_procs[bid].remove(proc)
 
 
 class popen_output(str):
